requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.7.4
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
supabase==2.0.2
email-validator==2.1.0
python-dotenv==1.0.0
openai==1.12.0
requests==2.31.0